requirements.txt
Flask==2.3.0
Flask-SQLAlchemy==3.1.0
Werkzeug==2.3.0
pandas
XlsxWriter==3.1.0
requests==2.31.0
mysql-connector-python==8.1.0
python-dotenv==1.0.0
email-validator==2.0.0